JSON-RPCとclose監視を分離
